add validation and file handling modules with utility functions for JPG import
