+++ file_handling/__init__.py
@@ -1,0 +1,4 @@
+from file_handling.discovery import find_jpg_files
+from file_handling.organization import get_destination_folder
+
+__all__ = ["find_jpg_files", "get_destination_folder"]
