+++ utils/__init__.py
@@ -1,6 +1,7 @@
 from utils.hashing.hashing import HashingUtils
 from utils.exif.exif import ExifUtils
 from utils.logs.logging_utils import LoggingUtils
+from utils.validation.validators import validate_directories
 
 # Create convenience functions that map to class methods
 get_hash = HashingUtils.get_hash
@@ -13,6 +14,7 @@ __all__ = [
     "get_date_taken",
     "compare_hashes",
     "get_hash",
+    "validate_directories",
     "HashingUtils",
     "ExifUtils",
     "LoggingUtils",
