+++ utils/validation/__init__.py
@@ -1,0 +1,3 @@
+from utils.validation.validators import validate_directories
+
+__all__ = ["validate_directories"]
